[skip ci] revert authenticate.[ch] split remove authenticate_extract.[ch]pp
